Removed Scodec.ByteVector from KeyInitializer
