+++ crypto/src/test/scala/co/topl/crypto/generation/KeyInitializerSpec.scala
@@ -9,7 +9,6 @@ import io.circe.{Decoder, HCursor}
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.propspec.AnyPropSpec
 import org.scalatestplus.scalacheck.ScalaCheckDrivenPropertyChecks
-import scodec.bits.ByteVector
 
 /**
  * test vectors adapted from multiple sources:
